pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "terminal-arcade"
version = "1.0.0"
description = "18 stunning terminal games, visualizations & tools in pure Python curses"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "wang2-lat"},
]
keywords = ["terminal", "arcade", "games", "curses", "ascii", "cli", "snake", "tetris", "matrix"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console :: Curses",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Topic :: Games/Entertainment",
    "Topic :: Games/Entertainment :: Arcade",
    "Topic :: Games/Entertainment :: Puzzle Games",
]

[project.optional-dependencies]
images = ["Pillow>=9.0"]

[project.urls]
Homepage = "https://github.com/wang2-lat/terminal-arcade"
Repository = "https://github.com/wang2-lat/terminal-arcade"

[project.scripts]
ta-snake = "terminal_arcade.snake:main_entry"
ta-tetris = "terminal_arcade.tetris:main_entry"
ta-matrix = "terminal_arcade.matrix:main_entry"
ta-globe = "terminal_arcade.globe:main_entry"
ta-fluid = "terminal_arcade.fluid:main_entry"
ta-life = "terminal_arcade.life:main_entry"
ta-maze = "terminal_arcade.maze:main_entry"
ta-particles = "terminal_arcade.particles:main_entry"
ta-music = "terminal_arcade.music_viz:main_entry"
ta-typing = "terminal_arcade.typing_test:main_entry"
ta-shooter = "terminal_arcade.space_shooter:main_entry"
ta-blackjack = "terminal_arcade.blackjack:main_entry"
ta-calculator = "terminal_arcade.calculator:main_entry"
ta-2048 = "terminal_arcade.game2048:main_entry"
ta-minesweeper = "terminal_arcade.minesweeper:main_entry"
ta-catpet = "terminal_arcade.catpet:main_entry"

[tool.setuptools.packages.find]
include = ["terminal_arcade*"]
